# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de la plateforme QUI PERD (défis de match, escrow, paiements Mobile Money).",
        "title": "QUI PERD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/paiements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "paiements"
                ],
                "summary": "Suivi des paiements (admin) — paginé, 10 par page, plus récents d'abord",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (défaut 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Éléments par page (1..100, défaut 10)",
                        "name": "taille",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "depot | retrait",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "en_attente | reussi | echoue | rembourse",
                        "name": "statut",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Page-paiements_Paiement"
                        }
                    }
                }
            }
        },
...
        "/paiements/depot": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "paiements"
                ],
                "summary": "Dépôt via LigdiCash/MoneyFusion",
                "responses": {}
            }
        },
        "/paiements/retrait": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "paiements"
                ],
                "summary": "Retrait vers Mobile Money",
                "responses": {}
            }
        },
        "/paiements/{id}/statut": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "paiements"
                ],
                "summary": "Validation/échec/remboursement manuel (admin)",
                "responses": {}
            }
        },
...
        "/portefeuille": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "portefeuille"
                ],
                "summary": "Solde du portefeuille",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/portefeuilles.Portefeuille"
                        }
                    }
                }
            }
        },
        "/portefeuille/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "portefeuille"
                ],
                "summary": "Historique des mouvements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/portefeuilles.TransactionPortefeuille"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "paiements.Paiement": {
            "type": "object",
            "properties": {
                "dateCreation": {
                    "type": "string"
                },
                "dateModification": {
                    "type": "string"
                },
                "devise": {
                    "type": "string"
                },
                "frais": {
                    "description": "frais de retrait (configurations_financieres.frais_retrait)",
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "montant": {
                    "type": "number"
                },
                "operateur": {
                    "type": "string"
                },
                "prestataire": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "referencePrestataire": {
                    "type": "string"
                },
                "statut": {
                    "type": "string"
                },
                "traite": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "utilisateurId": {
                    "type": "string"
                }
            }
        },
...
        "portefeuilles.Portefeuille": {
            "type": "object",
            "properties": {
                "dateCreation": {
                    "type": "string"
                },
                "dateModification": {
                    "type": "string"
                },
                "devise": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "soldeBloque": {
                    "type": "number"
                },
                "soldeDisponible": {
                    "type": "number"
                },
                "utilisateurId": {
                    "type": "string"
                }
            }
        },
        "portefeuilles.TransactionPortefeuille": {
            "type": "object",
            "properties": {
                "dateCreation": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "matchId": {
                    "type": "string"
                },
                "miseId": {
                    "type": "string"
                },
                "montant": {
                    "type": "number"
                },
                "portefeuilleId": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "statut": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "utils.Page-paiements_Paiement": {
            "type": "object",
            "properties": {
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/paiements.Paiement"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "taille": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}